Safari-611.1.10/LayoutTests/media/modern-media-controls/buttons-container/buttons-container-layout-expected.txt

Testing ButtonsContainer layout.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS container.children.length is 5
PASS tenPtWideButton.visible is true
PASS disabledButton.visible is false
PASS twentyPtWideButton.visible is true
PASS droppedButton.visible is false
PASS thirtyPtWideButton.visible is true
PASS tenPtWideButton.x is 20
PASS twentyPtWideButton.x is 40
PASS thirtyPtWideButton.x is 70
PASS container.width is 120
PASS successfullyParsed is true

TEST COMPLETE
